FAI-1312 - Move apprentice gov auth endpoint to new shared project
