+++ src/APIM.sln
@@ -490,6 +490,10 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.EmployerPR.Api.Unit
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SFA.DAS.EmployerPR.UnitTests", "EmployerPR\SFA.DAS.EmployerPR.UnitTests\SFA.DAS.EmployerPR.UnitTests.csproj", "{736F3868-DBC2-4687-8A10-A41A8E3089EC}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth", "Shared\SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth\SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth.csproj", "{36F403C1-1AB3-45DF-ACB2-6860BC78992D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth.UnitTests", "Shared\SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth.UnitTests\SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth.UnitTests.csproj", "{FDE38D44-8F72-4D83-8ED2-15855B2FF3B0}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -1280,6 +1284,14 @@ Global
 		{736F3868-DBC2-4687-8A10-A41A8E3089EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{736F3868-DBC2-4687-8A10-A41A8E3089EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{736F3868-DBC2-4687-8A10-A41A8E3089EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{36F403C1-1AB3-45DF-ACB2-6860BC78992D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{36F403C1-1AB3-45DF-ACB2-6860BC78992D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{36F403C1-1AB3-45DF-ACB2-6860BC78992D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{36F403C1-1AB3-45DF-ACB2-6860BC78992D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FDE38D44-8F72-4D83-8ED2-15855B2FF3B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FDE38D44-8F72-4D83-8ED2-15855B2FF3B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FDE38D44-8F72-4D83-8ED2-15855B2FF3B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FDE38D44-8F72-4D83-8ED2-15855B2FF3B0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -1481,6 +1493,8 @@ Global
 		{C05BB28B-0710-43F7-8D42-D69FBC635439} = {7717A4EF-73D9-413A-BD31-4B3D33557FBD}
 		{87C73F3B-DBFC-4448-A11A-4650B255768D} = {7717A4EF-73D9-413A-BD31-4B3D33557FBD}
 		{736F3868-DBC2-4687-8A10-A41A8E3089EC} = {7717A4EF-73D9-413A-BD31-4B3D33557FBD}
+		{36F403C1-1AB3-45DF-ACB2-6860BC78992D} = {9C327515-DF6A-4277-BF53-4C86D510C718}
+		{FDE38D44-8F72-4D83-8ED2-15855B2FF3B0} = {9C327515-DF6A-4277-BF53-4C86D510C718}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {2353A4D0-2859-4472-B2D4-BBD8087B4E22}
