+++ src/ApprenticePortal/SFA.DAS.ApprenticePortal.Api/SFA.DAS.ApprenticePortal.Api.csproj
@@ -47,6 +47,7 @@
         </Content>
     </ItemGroup>
     <ItemGroup>
+      <ProjectReference Include="..\..\Shared\SFA.DAS.SharedOuterApi.Apprentice.GovUk.Auth\SFA.DAS.SharedOuterApi.Apprentice.GovUk.Auth.csproj" />
       <ProjectReference Include="..\SFA.DAS.ApprenticePortal\SFA.DAS.ApprenticePortal.csproj" />
       <ProjectReference Include="..\..\Shared\SFA.DAS.SharedOuterApi\SFA.DAS.SharedOuterApi.csproj" />
     </ItemGroup>
