--- src/ApprenticePortal/SFA.DAS.ApprenticePortal/Application/Commands/ApprenticeAccounts/UpsertApprenticeCommandResult.cs
@@ -1,8 +1,0 @@
-using SFA.DAS.ApprenticePortal.Models;
-
-namespace SFA.DAS.ApprenticePortal.Application.Commands.ApprenticeAccounts;
-
-public class UpsertApprenticeCommandResult
-{
-    public Apprentice Apprentice { get; set; }
-}
