+++ src/Shared/SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth/Application/Commands/UpsertApprenticeCommandResult.cs
@@ -1,0 +1,6 @@
+namespace SFA.DAS.SharedOuterApi.Apprentice.GovUK.Auth.Application.Commands;
+
+public class UpsertApprenticeCommandResult
+{
+    public Models.Apprentice Apprentice { get; set; }
+}
